Start using dataset_info and add NDVI
This refactors things to pass around the dataset_info object instead of the individual components of it
to various functions which simplifies the code. We also added an NDVI channel which might improve
accuracy. This channel is baked into the arrays that are generated by the preprocess script. Unfortunately,
this increases the size of the dataset, but it made implementation simpler. The alternative is to have "dynamic"
channels that are computed from the "base" layers stored on disk.