Refactor to support multiple data generators and datasets
Previously, the code was tied to the Potsdam dataset using a single possible
data generator. This refactoring makes it easier to use multiple datasets and
generators. Now there are two generators for Potsdam: one that reads raw
TIFF/JPG data, and another that uses numpy arrays that have been saved to disk.
The numpy version is about 20x faster, but requires a preprocessing step. There
is also a PotsdamDataset class that has general info useful across different data
generators for Potsdam.

* Update README.md
* Create DataGenerator class hierarchy
* Create PotsdamDataset that contains data manipulation utilities
* Make preprocess a method of data generators so each generator can have its own
way of preprocessing the data
* Add command line args to generators.py
* Make utility functions handle batches or images for convenience, and establish a
convention of passing images instead of batches to functions that can only operate on a
single image at a time.