Fix problems with ISPRS preprocess methods
Previously, I forgot to switch to passing the options object to generator constructors in two of the
preprocess methods, so I fixed that. I also reverted to writing the channel stats at the end of the
preprocess method, since we need to ensure that the numpy files have already been generated. I also
changed the order of tasks so that preprocess happens before plot.