Make generator for aggregate jobs
Previously, we were using the generator of the first aggregated run for the aggregate run. This
is a problem when doing an agg_concat job, since each job has a different set of active_tags, and
we need to create the generator using the full set of active_tags, which is specified in the aggregate
job. In order to make this change, I needed to make it so that aggregate jobs have all the options necessary
to create a generator, so now aggregate jobs options files have that. Also, we now only download
aggregate run files if they aren't already on the file system to save time.