Improve check for dataset download
Previously, we checked to see if a dataset was downloaded by checking to see if
the dataset directory exists. This breaks if the directory was created, but the
files haven't finished downloading and extracting. Our new approach is to check
a done.txt file to see if it contains the list of files in the dataset. This
file is written to when download/extract is complete. Simply checking for the
existence of the files would not work since some of them are zip files which are
deleted after extract, and whose names might not map in a straightforward way
onto the extracted files.