Make using the "best" model an option
If "flying blind" and using no or little validation data, it makes sense to use the model found in the
final epoch instead of the "best" according to the validation score. Also, this fixes a bug wherein
the model from the final epochs was always used after the train_model task completed.