Add scripts to predict on large images with many objects
The strategy is to slide a window over the large image and generate a
directory full of window images. Then, we run the usual predict script over
that directory of images. Finally, we aggregate the predictions on the windows,
taking into account the offset of each window within the larger image.