Generate GeoJSON for output of aggregate predictions
Generate GeoJSON as output of aggregate predictions script, and use
the CRS and transform if input is GeoTIFF file. This also fixes the
rectangle grouping process so that instead of returning the cluster
centers, we return the original rectangles that are closest to the
cluster centers along with their scores.