Add --save-temp option
This adds an option so that scripts can put their temp files
into a directory at a known location and not delete that
directory on exit. Otherwise, temporary directories are put at
"random" locations and are deleted on exit, which is necessary
to avoid collisions when simultaneously running several instances of
the same script on a machine.