Update object detection docs with new test workflow
This workflow runs faster and makes it easier to see if the system works or not.
It's sort of a  poor man's integration test. This uses data from a new
repo https://github.com/azavea/raster-vision-data/releases which allows us to
distribute large files without including them in this repo.