Add ClassificationGeoJSONFile labels and related functionality
This represents classification data using a spatial grid where a class (or None)
is associated with each grid cell. We do not assume the GeoJSON file contains a
grid though. Instead we infer the grid from whatever polygons are in the file
along with options that determine how to infer the values of the grid cells.