Remove S3 windowed reads
Previously, we were not downloading GeoTIFFs from S3 in download_if_needed and
were letting Rasterio do windowed reads off of S3. This is probably slowing
things down since we are usually reading windows that cover the whole image.
So, I'm removing this functionality. In the future, we may want  to add this
back, but only when we're reading small regions of  the GeoTIFF and/or it's a
COG.