Add nodata data augmentation for object detection
This adds squares filled with zeros to negative chips in order to better
handle making predictions on scenes with NODATA values along the  edges in an
irregular pattern. It only works on negative chips since blacking  out random
parts of positive chips could mask objects, which would be misleading to the
training algorithm. This only works for object detection since there is no
notion of negative chips for the other tasks.