Start Again
```
lossyrob commented 14 hours ago
The dry run option seems pretty pervasive in the command configs,
runners, etc. I wonder if there is a less invasive set of changes that
gets to the same results, but uses a top level runner to do the work?
i.e. create a DryRunRunner that "runs" the experiment by describing the
configuration for commands it would run if it were a real runner? And
then the skipped commands could be handled by the top level
ExperimentRunner, or else delegated to inherited members, in which case
the DryRunRunner would report on the skipped commands. What do you
think?
```

...

...

...