Represent segmentation labels consistently
Previously, segmentation labels were sometimes returned as numpy arrays
and sometimes as SemanticSegmentationLabels objects. Somehow this worked
up until now, but it was inconsistent and confusing and also broke the integration
test. So this refactors so that labels are always returned as
SemanticSegmentationLabels objects and adds to_ and from_array methods to
transform to and from arrays when needed.