S3FileSystem Unit Tests
sync_(to|from)_dir Not Testable

Both of those functions rely on the AWS CLI which (apparently) uses the
Python `requests` library.  An error message from within the `requests`
library is thrown when one tries to connect to a mock bucket while
`moto` is active.

There seems to be a fundamental incompatibility (causes by the
`requests` library) between `moto` and the AWS CLI.

...

...