Make sure each plugin imports rv2.pipeline first thing
This is needed so that the registry is setup before anything else happens.
It's only needed if you are importing something in a plugin that is not being
run via the pipeline.runner. I doubt this will happen much, but I'm adding
this to be safe. It seems a little odd that we need to do this and could
 lead to bugs if people forget to do it, but this is the best I've got right now.