Change chipping options to handle small scenes in Spacenet Vegas
If we use a chip size of 300 with 650x650 scenes, the bottom and right edges
will tend to do badly, and because the scenes are so small these errors
will have a big impact on the score. So we change the training and predict chip size
to evenly divide the scene size. This was true for the bottom edge in
rv1, and for some reason it started happening on the right edge in rv2.
That's strange, but since the broader issue affects both versions
(and any approach using a sliding window for that matter), this seems to be an
appropriate fix until we have a better prediction window generation strategy.