Update all relevant saved URIs in config before instantiating Pipeline
This is to fix the ability to use the Potsdam model bundle after the changes
in https://github.com/azavea/raster-vision/pull/972. That PR made it so that
instantiating the SemanticSegmentationPipeline instantiates the first scene.
Before this commit, the Predictor instantiated the Pipeline before updating
the raster source and transformer URIs, which was a problem because the
config in the model bundle contains URIs which point to S3 (or wherever the
models were trained).