Downgrade Sphinx to 3.5
The 4.0 version breaks the styling and it doesn't seem easy to fix
See the changelog for Sphinx 4.0:
"Support docutils-0.17. Please notice it changes the output of HTML builder. Some themes do not support it, and you need to update your custom CSS to upgrade it."
See https://github.com/sphinx-doc/sphinx/issues/9001