fix to ensure registry is set up correctly during unit tests
This is meant to fix a problem that occurred when running a unit test for saving and loading Learner's model bundle. Specifically, when `from_model_bundle()` called `upgrade_config(config_dict)`, it would throw a `KeyError` for "rastervision.core" due to "rastervision.core" being absent from the registry.

This was traced down to be due to the following workflow:
- `unittest discover` runs all discovered test files in alphabetical order
- the first file happens to be `tests/core/data/crs_transformer/test_rasterio_crs_transformer.py`
- the line `from rastervision.core.data.crs_transformer import RasterioCRSTransformer` in this file is run
- this causes core's `__init__.py`'s `import rastervision.pipeline` line to be run
- this causes the registry to be initialized
- the registry attempts to find all the plugins and load their `register_plugin()` functions
- (not 100% certain about this part but it seems like) when the registry tries to import core, it results in a circular import. At this point, core is already in `sys.modules` but has not been fully executed yet. Which means its `register_plugin()` hasn't been defined yet! Some context: https://stackoverflow.com/a/744403/5908685 .
- the registry fails to find core's `register_plugin()` and therefore fails to register it

This commit adds `import rastervision.pipeline` to `tests/__init__.py` and ensures that it gets executed before any of the test files by specifying the -t option in the unittest command.