Avoid yield in try block
If you have yield inside a context manager block and don't iterate through all the items,
it is difficult to predict when the exit method of the context manager will be called.
This led to a bug where torch was inadvertently in inference mode until the garbage
collector happened to delete the generator object. This is fixed by removing the
use of the context manager here which is unnecessary since it is called in the
predict method. We also get rid of the call to self.model.train because it will only
get called if all the items are iterated through, and it's redundant since the
model is explicitly put back into training mode at the start of the train method.