Add --arm64 flag to build and run ARM64 image
The name of the ARM64 image has a -arm64 suffix.
Keeping the current name of the AMD64 image minimizes
disruption for most users. Once we start publishing the
two images to Quay in the future, we should name them
both with a suffix. It is important for them to have
distinct names though for users (such as me) who need to
switch between both types of image.