determine num_channels and dtype statically w/o reading a chip
The num_channels and dtype of a RasterSource can be different from the underlying data source because of RasterTransformers. Previously, these could only be determined dynamically at runtime by reading a 1x1 chip and checking its shape and dtype. This can be very slow when initializing many RasterSources or when initializing a RasterSources with many bands.

This change adds get_out_channels() and get_out_dtype() methods to each RasterTransformer, allowing num_channels and dtype to be determined statically without reading any data.