handle non-deterministic pyproj failure
For some transformations, pyproj attempts to download transformation
grids from the internet for improved accuracy when
Transformer.transform() is called. If it fails to connect to the
internet, it silently returns (inf, inf) and silently modifies its
behavior to not access the internet on subsequent calls, causing
them to succeed (though possibly with a loss of accuracy). See
https://github.com/pyproj4/pyproj/issues/705 for details.

This workaround forces an error to be raised by setting
errcheck=True and ignoring the first error.